matgo/matgo-crawler/src/requirements.txt
selenium==4.25.0
pymongo==4.10.1
fake-useragent==1.5.1
beautifulsoup4==4.12.3
python-dotenv==1.0.1
urllib3==2.2.3
torch==2.5.1
requests==2.32.3
transformers==4.47.0
pillow==11.0.0